renamed diseases tab filter components
