+++ client/client/app/components/ngbTargetPanel/ngbDiseasesTab/ngbDiseasesTargetsPanel/ngbDiseasesTargetsTableFilter/index.js
@@ -5,8 +5,8 @@ import './ngbDiseasesTargetsTableFilter.scss';
 import component from './ngbDiseasesTargetsTableFilter.component';
 import controller from './ngbDiseasesTargetsTableFilter.controller';
 
-import filterInput from './ngbTargetsFilterInput';
-import filterList from './ngbTargetsFilterList';
+import filterInput from './ngbDiseasesTargetsFilterInput';
+import filterList from './ngbDiseasesTargetsFilterList';
 
 export default angular.module('ngbDiseasesTargetsTableFilter', [filterList, filterInput])
     .controller(controller.UID, controller)
