+++ client/client/app/components/ngbTargetPanel/ngbDiseasesTab/ngbDiseasesTargetsPanel/ngbDiseasesTargetsTableFilter/ngbDiseasesTargetsFilterInput/ngbDiseasesTargetsFilterInput.component.js
@@ -1,0 +1,9 @@
+import controller from './ngbDiseasesTargetsFilterInput.controller';
+
+export default  {
+    bindings: {
+        column: '<'
+    },
+    controller: controller.UID,
+    template: require('./ngbDiseasesTargetsFilterInput.tpl.html')
+};
