+++ client/client/app/components/ngbTargetPanel/ngbDiseasesTab/ngbDiseasesTargetsPanel/ngbDiseasesTargetsTableFilter/ngbDiseasesTargetsFilterInput/ngbDiseasesTargetsFilterInput.tpl.html
@@ -4,3 +4,4 @@
        ng-model="$ctrl.value"
        style="margin-bottom: 1px"
        type="text">
+       <span>hello</span>
