+++ client/client/app/components/ngbTargetPanel/ngbDiseasesTab/ngbDiseasesTargetsPanel/ngbDiseasesTargetsTableFilter/ngbDiseasesTargetsFilterList/ngbDiseasesTargetsFilterList.component.js
@@ -1,0 +1,9 @@
+import controller from './ngbDiseasesTargetsFilterList.controller';
+
+export default  {
+    bindings: {
+        column: '<'
+    },
+    controller: controller.UID,
+    template: require('./ngbDiseasesTargetsFilterList.tpl.html')
+};
