+++ client/client/app/components/ngbTargetPanel/ngbDiseasesTab/ngbDiseasesTargetsPanel/ngbDiseasesTargetsTableFilter/ngbDiseasesTargetsFilterList/ngbDiseasesTargetsFilterList.scss
