+++ client/client/app/components/ngbTargetPanel/ngbDiseasesTab/ngbDiseasesTargetsPanel/ngbDiseasesTargetsTableFilter/ngbDiseasesTargetsTableFilter.tpl.html
@@ -1,6 +1,6 @@
-<ngb-targets-filter-input column="$ctrl.column"
+<ngb-diseases-targets-filter-input column="$ctrl.column"
                           ng-if="$ctrl.isString">
-</ngb-targets-filter-input>
+</ngb-diseases-targets-filter-input>
 <ngb-diseases-targets-filter-list column="$ctrl.column"
                                   ng-if="$ctrl.isList">
 </ngb-diseases-targets-filter-list>
