--- client/client/app/components/ngbTargetPanel/ngbDiseasesTab/ngbDiseasesTargetsPanel/ngbDiseasesTargetsTableFilter/ngbTargetsFilterInput/index.js
@@ -1,9 +1,0 @@
-import angular from 'angular';
-
-import component from './ngbTargetsFilterInput.component';
-import controller from './ngbTargetsFilterInput.controller';
-
-export default angular.module('ngbTargetsFilterInput', [])
-    .controller(controller.UID, controller)
-    .component('ngbTargetsFilterInput', component)
-    .name;
