--- client/client/app/components/ngbTargetPanel/ngbDiseasesTab/ngbDiseasesTargetsPanel/ngbDiseasesTargetsTableFilter/ngbTargetsFilterInput/ngbTargetsFilterInput.component.js
@@ -1,9 +1,0 @@
-import controller from './ngbTargetsFilterInput.controller';
-
-export default  {
-    bindings: {
-        column: '<'
-    },
-    controller: controller.UID,
-    template: require('./ngbTargetsFilterInput.tpl.html')
-};
